requirements.txt
python-telegram-bot==20.3
mysql-connector-python
playwright
cryptography
fastapi
uvicorn
jinja2
python-dotenv
aiohttp
requests